rename legalInformation to LegalInformation
