+++ frontend/src/components/auth/LegalInformation.vue
