+++ frontend/src/pages/account/settings.vue
@@ -105,9 +105,9 @@
 <script setup lang="ts">
 import type { UserAttributes } from '@supabase/supabase-js';
 
-import LegalInformation from '@/components/auth/legalInformation.vue';
 import { isInvalidEmail } from '@/utils/email';
 import { isInvalidPassword } from '@/utils/password';
+import LegalInformation from '~/components/auth/LegalInformation.vue';
 import type { Profile } from '~/types/profile';
 
 const { t } = useI18n({
